pyproject.toml
[project]
name = "livraria"
version = "1.0.0"
description = "Uma API REST de uma livraria feita com Django e Django REST Framework."
authors = [{ name = "Marco André Mendes", email = "marcoandre@gmail.com" }]
dependencies = [
  'dj-database-url>=2.1.0',
  'django-cloudinary-storage>=0.3.0',
  'django-cors-headers>=4.3.0',
  'django-extensions>=3.2.3',
  'django-filter>=23.3',
  "django>=5.1.3",
  'djangorestframework>=3.14.0',
  'drf-spectacular>=0.26.5',
  'gunicorn>=21.2.0',
  #'mysqlclient>=2.2.4',
  'passage-identity>=2.5.1',
  'Pillow>=10.3.0',
  'pre-commit>=3.5.0',
  'psycopg2-binary>=2.9.9',
  'pydotplus>=2.0.2',
  'python-dotenv>=1.0.0',
  'python-magic-bin>=0.4.14; sys_platform == "win32"',
  'python-magic>=0.4.27; sys_platform == "linux"',
  'setuptools>=68.2.2',
  'uvicorn>=0.29.0',
  'whitenoise[brotli]>=6.6.0',
]
requires-python = ">=3.10"
license = { text = "MIT" }

[tool.pdm]
distribution = false

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"

[tool.ruff]
line-length = 120
extend-exclude = ['migrations', '__init__.py', 'manage.py']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ["PLR6301"]

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver 0.0.0.0:19003"
createsuperuser = "python manage.py createsuperuser"
pre_migrate = "python manage.py makemigrations"
migrate = "python manage.py migrate"
post_migrate = "python manage.py graph_models -S -g -o core.png core"
shell = "python manage.py shell"
shellp = "python manage.py shell_plus"
test = "python manage.py test"
loaddata = "python manage.py loaddata core.json"
dumpdata = "python manage.py dumpdata --indent 2"
cria_api = "python ./scripts/cria_api.py {args}"
check = "ruff check"
pre_format = "ruff check --fix"
format = "ruff format"

[dependency-groups]
dev = [
  "pylint>=3.0.2",
  "ipython>=8.26.0",
  "ruff>=0.11.2",
]
